webassembly code

+++ floats.c
@@ -1,4 +1,4 @@
 
-int toInteger(float a){
+int toInteger(int a){
 	return (int)a;
 }
